backend/requirements.txt
APIFlask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
alembic
python-dotenv
flask-cors
ruff
flask-jwt-extended

# HTTP Requests
requests

# Testing
pytest
pytest-flask
factory-boy
pytest-cov

# Async Task Queue
celery
redis

# PDF Generation
WeasyPrint

# Excel Generation
openpyxl
